+++ Sources/CorePayments/Networking/GraphQL/GraphQLQuery.swift
@@ -1,7 +1,9 @@
 import Foundation
 
-protocol GraphQLQuery {
+//// :nodoc: Protocol for requests made to GraphQL endpoints
+public protocol GraphQLQuery {
+    associatedtype VariablesType: Codable
     var query: String { get }
-    var variables: [String: Any] { get }
+    var variables: VariablesType? { get }
     func requestBody() throws -> Data
 }