+++ Sources/CorePayments/Networking/GraphQL/GraphQLQueryResponse.swift
@@ -1,4 +1,5 @@
-struct GraphQLQueryResponse<T: Codable>: Codable {
+//// :nodoc: Struct to handle nested response from GraphQL response
+public struct GraphQLQueryResponse<T: Codable>: Codable {
 
-    let data: T?
+    public let data: T?
 }