+++ UnitTests/CardPaymentsTests/CardClient_Tests.swift
@@ -25,6 +25,8 @@ class CardClient_Tests: XCTestCase {
     var mockAPIClient: MockAPIClient!
     var cardClient: CardClient!
     var cardRequest: CardRequest!
+    var mockGraphQLClient: MockGraphQLClient!
+    var mockCardVaultDelegate: MockCardVaultDelegate!
     // swiftlint:enable implicitly_unwrapped_optional
 
     // MARK: - Test lifecycle
@@ -34,12 +36,82 @@ class CardClient_Tests: XCTestCase {
         config = CoreConfig(clientID: mockClientID, environment: .sandbox)
         mockAPIClient = MockAPIClient(coreConfig: config)
         cardRequest = CardRequest(orderID: "testOrderId", card: card)
-
+        mockGraphQLClient = MockGraphQLClient(environment: .sandbox)
+        
         cardClient = CardClient(
             config: config,
             apiClient: mockAPIClient,
-            webAuthenticationSession: mockWebAuthSession
+            webAuthenticationSession: mockWebAuthSession,
+            graphQLClient: mockGraphQLClient
         )
+    }
+    
+    // MARK: - vault() tests
+
+    func testVault_withValidResponse_returnsSuccess() {
+        let setupTokenID = "testToken1"
+        let vaultStatus = "APPROVED"
+        let vaultRequest = CardVaultRequest(card: card, setupTokenID: setupTokenID)
+        let updateSetupTokenResponse = UpdateSetupTokenResponse(
+            updateVaultSetupToken: TokenDetails(id: setupTokenID, status: vaultStatus, links: [TokenDetails.Link(rel: "df", href: "h")])
+        )
+        mockGraphQLClient.mockSuccessResponse = GraphQLQueryResponse(data: updateSetupTokenResponse)
+        
+        let expectation = expectation(description: "vault completed")
+        let cardVaultDelegate = MockCardVaultDelegate(success: {_, result in
+            XCTAssertEqual(result.setupTokenID, setupTokenID)
+            XCTAssertEqual(result.status, vaultStatus)
+            expectation.fulfill()
+        }, error: {_, _ in
+            XCTFail("Invoked error() callback. Should invoke success().")
+        })
+        cardClient.vaultDelegate = cardVaultDelegate
+        cardClient.vault(vaultRequest)
+        
+        waitForExpectations(timeout: 10)
+    }
+    
+    func testVault_withNoData_ReturnsError() {
+        let setupTokenID = "testToken1"
+        let vaultRequest = CardVaultRequest(card: card, setupTokenID: setupTokenID)
+       
+        mockGraphQLClient.mockSuccessResponse = GraphQLQueryResponse(data: nil)
+        
+        let expectation = expectation(description: "vault completed")
+        let cardVaultDelegate = MockCardVaultDelegate(success: {_, _ in
+            XCTFail("Invoked success() callback. Should invoke error().")
+        }, error: {_, error in
+            XCTAssertEqual(error.domain, CardClientError.domain)
+            XCTAssertEqual(error.code, CardClientError.Code.noVaultTokenDataError.rawValue)
+            XCTAssertEqual(error.localizedDescription, "No data was returned from update setup token service.")
+            expectation.fulfill()
+        })
+        cardClient.vaultDelegate = cardVaultDelegate
+        cardClient.vault(vaultRequest)
+        
+        waitForExpectations(timeout: 10)
+    }
+
+    func testVault_whenGraphQLCallFails_returnsError() {
+        let setupTokenID = "testToken1"
+        let vaultRequest = CardVaultRequest(card: card, setupTokenID: setupTokenID)
+       
+        mockGraphQLClient.mockErrorResponse = GraphQLError(
+            message: "thee was an error fetching data from GraphQL endpoint", extensions: nil
+        )
+        let expectation = expectation(description: "vault completed")
+        let cardVaultDelegate = MockCardVaultDelegate(success: {_, _ in
+            XCTFail("Invoked success() callback. Should invoke error().")
+        }, error: {_, error in
+            XCTAssertEqual(error.domain, CardClientError.domain)
+            XCTAssertEqual(error.code, CardClientError.Code.vaultTokenError.rawValue)
+            XCTAssertEqual(error.localizedDescription, "An error occurred while vaulting a card.")
+            expectation.fulfill()
+        })
+        cardClient.vaultDelegate = cardVaultDelegate
+        cardClient.vault(vaultRequest)
+        
+        waitForExpectations(timeout: 10)
     }
 
     // MARK: - approveOrder() tests